+++ src/reactive/array.ts
@@ -281,7 +281,6 @@ export abstract class ReactiveArrayWrapper extends Array {
 		return reactive(super.toSorted(wrappedCompare))
 	}
 
-	// biome-ignore lint/complexity/noArguments: Needed to distinguish fn() from fn(undefined)
 	toSpliced(start: number, deleteCount?: number, ...items: any[]): any {
 		if (arguments.length > 2)
 			return reactive(super.toSpliced(start, deleteCount!, ...items.map(unwrap)))